backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1

# Supabase (using REST API directly - no version conflicts)
requests==2.31.0

# ML & AI
# PyTorch CUDA wheels are hosted on the PyTorch index (not PyPI).
--index-url https://download.pytorch.org/whl/cu118
--extra-index-url https://pypi.org/simple
torch==2.1.2+cu118
torchvision==0.16.2+cu118
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.0
Pillow==10.2.0
opencv-python==4.9.0.80

# Hugging Face (disease detection)
transformers==4.36.0

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2

# External APIs
aiohttp==3.9.1
httpx>=0.26.0

# Rate Limiting
slowapi>=0.1.9

# Retry Logic
tenacity>=8.2.0

# Text-to-Speech (9 Indian languages via Microsoft Edge TTS)
edge-tts>=6.1.9

# Post-Quantum Cryptography
pqcrypto>=0.4.0
pycryptodome==3.19.0
# Optional NeoShield lattice implementation (preferred over fallback Falcon in NeoShield module)
dilithium-py>=0.1.3

# Analytics
duckdb>=0.10.0

# Satellite / Geospatial
sentinelhub>=3.9.0
rasterio>=1.3.0
Shapely>=2.0.0
pyproj>=3.5.0

# New analytics / ML
scikit-image>=0.21.0
scipy>=1.11.0
joblib>=1.3.0

# WhatsApp / Twilio (already present - ensure version)
twilio==8.11.1

# Web3 (lightweight - for parametric insurance ledger)
web3>=6.0.0

# Utils
python-dotenv==1.0.0
matplotlib>=3.8.0

# Voice Assistant (optional - requires manual install)
# openai-whisper  # Run: pip install -U openai-whisper
# edge-tts  # Run: pip install edge-tts

# Image/GPS processing
exifread>=3.0.0

# WhatsApp integration (Twilio already included above)
